Remove mapping table and create ContentProviderBase
